2019-07-04 added tabgroup to the ui prototyped the fields for each tab
2019-07-05
added code to fill the company and contact tabs with data.
added code to fill the action item tab
added code to fill the contact log tab
pulled the company list box out of the tab group so that it is always visible

2019-07-06
added code to save changes to the tables
added code to create companies - have to manually create related records
in the other tables.

2019-07-10
fixed bug - reading rows where the rowid was greater than 9 was crashing
the script.
fixed bug - no longer crashes after creating a new company